Update README.md
So here the stuff im trying to work on, Im in the process of learning scripts...
theres a couple of syntaxes in here i know arnt correct ,and it tells my name (if you can read the code) but this is just for me to get to know github and practice a little scripting as well. Any usefull thoughts and ideas are welcome. Please dont be condesing as i learn by makeing mistakes just tell me whats worng so i can fix it and learn not to do it that way agian next time